pyproject.toml
[tool.poetry]
name = "scipylat-discord-bot"
version = "0.1.0"
description = ""
authors = ["Arthur Alvim <afmalvim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.7.3"
sentry-sdk = "^1.4.3"
python-decouple = "^3.5"
pytz = "^2021.3"
gspread = "^4.0.1"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
